CmdOpen1: compute unite-panel y inner term in single precision
Target computes (animEntry->height * scale) - baseEntry->height with
fsubs/fnmsubs (single precision), then the outer *0.5 + baseEntry->y -
offset in double. Our source forced the inner term to double via
static_cast<f64>; changing those casts to f32 makes MWCC emit the
single-precision fsubs/fnmsubs chain the target uses.
CmdOpen1 67.75 -> 71.69, unit 72.53 -> 72.69.

+++ src/menu_cmd.cpp
@@ -2641,8 +2641,8 @@ unsigned int CMenuPcs::CmdOpen1()
 		animEntry->scale = static_cast<f32>(panelScale);
 		animEntry->width = 0xC0;
 		animEntry->height = 0x40;
-		animEntry->y = static_cast<s16>(((-((static_cast<f64>(animEntry->height) * animEntry->scale) -
-		                                    static_cast<f64>(baseEntry->height)) *
+		animEntry->y = static_cast<s16>(((-((static_cast<f32>(animEntry->height) * animEntry->scale) -
+		                                    static_cast<f32>(baseEntry->height)) *
 		                                   kCmdMenuHalfD) +
 		                                  static_cast<f64>(baseEntry->y)) -
 		                                 kCmdMenuUnitePanelYOffsetD);
